bibledotdev/requirements.txt
chromadb
orjson
opentelemetry-exporter-otlp-proto-http
reflex
rich
typer
traceloop-sdk
uuid7
